# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/session:
    get:
      operationId: getAuthSession
      tags: [auth]
      summary: Get the signed-in profile and staff role
      responses:
        "200":
          description: Active signed-in session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401": { description: Authentication required }
  /auth/register:
    post:
      operationId: registerAccount
      tags: [auth]
      summary: Create a customer or driver account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: New signed-in account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400": { description: Invalid registration details }
        "409": { description: An account already uses this email address }
  /auth/password-reset:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset email
      description: Always accepts a validly shaped request without revealing whether the email belongs to an active account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "202":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResetAccepted"
  /auth/password-reset/confirm:
    post:
      operationId: confirmPasswordReset
      tags: [auth]
      summary: Consume a password reset link and set a new password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmation"
      responses:
        "200":
          description: Password changed and a new session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400": { description: Malformed reset link or invalid password }
        "410": { description: Expired or already-used reset link }
...
components:
...
  schemas:
...
    AuthSession:
      type: object
      required: [profile]
      properties:
        profile:
          type: object
          required: [id, firstName, lastName, email, emailVerified, phoneVerified, role]
          properties:
            id: { type: string }
            firstName: { type: string }
            lastName: { type: string }
            email: { type: string }
            phone: { type: string, nullable: true }
            emailVerified: { type: boolean }
            phoneVerified: { type: boolean }
            role: { type: string, enum: [customer, driver, admin, dispatcher, support] }
    RegistrationInput:
      oneOf:
        - $ref: "#/components/schemas/CustomerRegistrationInput"
        - $ref: "#/components/schemas/DriverRegistrationInput"
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          maxLength: 254
          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
    PasswordResetAccepted:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean, const: true }
    PasswordResetConfirmation:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 1, maxLength: 256 }
        password: { type: string, minLength: 10, maxLength: 256 }
    CustomerRegistrationInput:
      type: object
      required: [firstName, lastName, email, password, phone]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        email:
          type: string
          maxLength: 254
          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
        password: { type: string, minLength: 10, maxLength: 256 }
        phone:
          type: string
          description: Include 10 to 15 digits; separators are allowed.
          maxLength: 32
          pattern: '^(?=(?:\D*\d){10,15}\D*$)[+().\s\d-]+$'
        accountType:
          type: string
          enum: [customer]
          default: customer
    DriverRegistrationInput:
      type: object
      required: [firstName, lastName, email, password, accountType]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 80 }
        lastName: { type: string, minLength: 1, maxLength: 80 }
        email:
          type: string
          maxLength: 254
          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
        password: { type: string, minLength: 10, maxLength: 256 }
        phone:
          type: string
          nullable: true
          maxLength: 32
        accountType:
          type: string
          enum: [driver]
...